example-base-03/backend/chat-with-my-bot-python-backend/requirements.txt
flask
mysql-connector-python
spacy